蜗速环岛

+++ project/code/judgement.h
@@ -11,6 +11,7 @@
 #include "zf_common_headfile.h"
 #include "Read_ADC.h"
 
+
 // µÀÂ·×´Ì¬Ã¶¾Ù
 typedef enum
 {
@@ -18,10 +19,12 @@ typedef enum
     Status_Circle = 1
 }RaceStatus;
 
+#include "servo.h"
+
 void judgement();
 
 extern uint8 left_circle_flag, right_circle_flag, cross_flag;
-extern float circle_threshold, cross_threshold;
+extern uint8 circle_threshold, cross_threshold;
 extern uint8 circle_status;
 
 #endif /* JUDGEMENT_H_ */
